Minor Service Adjustments

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/mapper/FileMapper.java
@@ -24,6 +24,6 @@ public interface FileMapper {
     Integer insertFileData(Files files);
 
     @Delete("DELETE from FILES WHERE fileid=#{fileId}")
-    Integer deleteById(long id);
+    Integer deleteById(Integer id);
 
 }
