+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/services/home/files/FileService.java
@@ -15,5 +15,5 @@ public interface FileService {
 
     Integer insertFiles(MultipartFile file, Integer userId);
 
-    Integer deleteFiles(long id);
+    Integer deleteFiles(Integer id);
 }
