+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/services/home/files/FileServiceImpl.java
@@ -51,7 +51,7 @@ public class FileServiceImpl implements FileService {
         }
     }
     @Override
-    public Integer deleteFiles(long id) {
+    public Integer deleteFiles(Integer id) {
         return fileMapper.deleteById(id);
     }
 }
